rename confusingly named component
