+++ src/components/Location.vue
