+++ src/views/Home.vue
@@ -12,7 +12,7 @@
 // @ is an alias to /src
 import Landing from '@/components/Landing.vue'
 import Mission from '@/components/Mission.vue'
-import Meetings from '@/components/Meetings.vue'
+import Meetings from '@/components/Location.vue'
 import Involved from '@/components/Involved.vue'
 import SocialMedia from '@/components/SocialMedia.vue'
